% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_versions.R
\name{get_versions}
\alias{get_versions}
\title{Get OncoTree versions}
\usage{
get_versions()
}
\value{
A \link[tibble:tibble-package]{tibble} of four variables:
\describe{
\item{\code{oncotree_version}}{OncoTree tumor classification system version.}
\item{\code{description}}{OncoTree release description.}
\item{\code{visible}}{A logical indicating whether this OncoTree version is visible, i.e. a forefront option at the website.}
\item{\code{release_date}}{OncoTree release date.}
}
}
\description{
Get OncoTree versions
}
\examples{
try(get_versions())
}
