% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_to_character.R
\name{logical_to_character}
\alias{logical_to_character}
\title{Converts a logical vector to character}
\usage{
logical_to_character(
  x,
  format = c("lowercase", "uppercase", "titlecase", "integer")
)
}
\arguments{
\item{x}{A logical vector.}

\item{format}{Controls how the text is generated.}
}
\value{
A character vector.
}
\description{
This function converts a logical vector to a character vector allowing to
control the case and even to convert to integer 0 or 1 if desired.
}
\keyword{internal}
