% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mskcc.oncotree-package.R
\docType{package}
\name{mskcc.oncotree-package}
\alias{mskcc.oncotree}
\alias{mskcc.oncotree-package}
\title{mskcc.oncotree: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
