% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_nci_thesaurus.R
\name{open_in_nci_thesaurus}
\alias{open_in_nci_thesaurus}
\title{Browse the NCIt}
\usage{
open_in_nci_thesaurus(nci_code)
}
\arguments{
\item{nci_code}{A character vector of NCI codes.}
}
\value{
Run for its side effect.
}
\description{
Opens the web browser at NCI Thesaurus for the entries provided as NCI codes.
}
\examples{
\dontrun{
open_in_nci_thesaurus('C3107')
}
}
